Edit: moved level to base.py
