+++ src/finch/tensor/level/sparse_coo_level.py
@@ -4,7 +4,7 @@ from typing import Any, cast
 import numpy as np
 
 from finch.algebra import FType, ImmutableStructFType, TupleFType, ftype, ftypes
-from finch.tensor.fiber_tensor import Level, LevelFType
+from finch.tensor.base import Level, LevelFType
 
 
 class SparseCOOLevelFType(ImmutableStructFType, LevelFType):
